OGLS.Engine.Bindings

Description

Documentation

data State

Constructors

State

Fields

angle :: IORef GLdouble

Angle of inclination of particular model

scaleto :: IORef GLfloat

I use scaling model instead of moving camera aroind it

delta :: IORef GLfloat

Needed for calculating inertia

inertia :: IORef (Vector3 GLfloat)

Initial inertia of the given model

inertiaOld :: IORef (Vector3 GLfloat)

Buffer for proper calculation

diff :: IORef (Vector3 GLfloat)

Difference of inclination while rotating model

shouldRotate :: IORef Bool

Rotation toggle (on/off)

useShaders :: IORef Bool

Shading toggle (on/off)

point :: IORef Int

Don't remember where I used it

modelsCycle :: IORef [IO ()]

Where models stored

bgCycle :: IORef [Color4 GLclampf]

Where backgrounds/skyboxes stored

matCycle :: IORef [Material]

Where materials stored

lastPosition :: IORef Position

I'm so fucking tired

lastIncr :: IORef (Vector3 GLfloat)

Last increment while rotating model

modifiers :: IORef Modifiers

Needed for proper keyboardMouse callback work

pos :: IORef (GLfloat, GLfloat)

I wanted to move model around the scene but it's unimplemented rn

data ShaderFilePath

Constructors

ShaderFilePath

Fields

vertexShaderFilePath :: FilePath

Path to .vert file

fragmentShaderFilePath :: FilePath

Path to .frag file

data Material

Constructors

Material

Fields

mName :: MaterialName

Name of material being used

mAmbient :: !(Color4 GLfloat)

Ambient map of the given material

mDiffuse :: !(Color4 GLfloat)

Diffuse map of the given material

mSpecular :: !(Color4 GLfloat)

Specular map of the given material

mEmission :: !(Color4 GLfloat)

Emission ratio of the given material

mShininess :: !GLfloat

Shininess ration of the given material

mRefraction :: !GLfloat

Refraction ratio of the given material

shaderFilePath :: ShaderFilePath

timerFrequency :: Timeout

toVertex :: Num a => (a, a) -> IO ()

showIO :: Show a => IO a -> IO ()

Shows contents of the given IO monad

getTime :: IO Double

Don't remember where I used it

prepareState :: IO State

Initialize state

nextBackground :: State -> IO ()

Toggles background color/skybox

scaleIncrement :: Float

scaleFactor :: Float

resetState :: State -> IO ()

Resets given state

runInertia :: State -> IO ()

Calculates inertia using given state

toggleRotation :: State -> IO ()

Toggles rotation for the given state

help :: Text

Outputs help to stdout

printHelp :: IO ()

Pointless trying to draw text on the window

force :: Strategy a -> Strategy [a]

Force evaluation of the given list force :: [a] -> () force (x:xs) = x pseq force xs force _ = ()